Update README.md
The data contains the geometry structure of all the configurations in the studied is stored in CONTCAR format, as well as the properties extracted from the DFT calculations carried out with VASP program. It is also includes the high-throughput vaspkit code (https://vaspkit.com/) related to thermodynamic results, the  results from the thermodynamical analysis performed with vaspkit is pointed out (data-Figure 9 and 10.xlsx).